Add tests for login-logout-login-external sequnce
